Make air movement feel less "floaty"
Movement while in the air is now accomplished through an accelerative force, as opposed to the instant velocity change used while on the ground. This gives the playere less control in the air and commits them more to the direction they were headed when they jumped, depending on the "Air Control" value.